mvp/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mvp"
version = "0.0.1"
description = "Proj_ongoing MVP: Beneish M-Score + Altman Z-Score vertical slice"
requires-python = ">=3.11,<3.12"
readme = "README.md"
authors = [{ name = "Sen Yang", email = "sy2576@stern.nyu.edu" }]

dependencies = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "pydantic>=2.6",
    "pyyaml>=6",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pytest>=8",
    "pytest-asyncio",
    "hypothesis>=6",
    "python-dateutil",
    "jsonschema>=4",
]

[project.scripts]
mvp = "mvp.cli.main:main"

[project.optional-dependencies]
# Layout-aware PDF parsing for academic papers. Heavy ONNX deps; excluded from the
# base install so the MVP can be set up on machines without ONNX-runtime support.
layout = ["pdf2zh"]

dev = [
    "pytest>=8",
    "pytest-asyncio",
    "hypothesis>=6",
    "mypy>=1.10",
    "ruff>=0.5",
]

# The working directory is `mvp/`, but the Python package is named `mvp` and we
# want `import mvp.lib.hashing` to resolve to `./lib/hashing.py`. Treat the
# current directory as the contents of the `mvp` package via package-dir.
[tool.setuptools]
packages = [
    "mvp",
    "mvp.lib",
    "mvp.ingestion",
    "mvp.store",
    "mvp.standardize",
    "mvp.engine",
    "mvp.skills",
    "mvp.skills.fundamental",
    "mvp.skills.interpretation",
    "mvp.skills.paper_derived",
    "mvp.skills.composite",
    "mvp.agents",
    "mvp.api",
    "mvp.cli",
    "mvp.scripts",
    "mvp.eval",
]
package-dir = { "mvp" = "." }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
